refactor: split semantic proof surface bands
Replace mixed semantic-proof roots with explicit fact, surface, and suite bands while keeping the public proof API stable.
